Make a successful post creation go back home
Before this change, if you submitted a job, the page would look
like it froze and that nothing happened. This might lead
people to click submit multiple times. Now we redirect
back to the home page, where a person will be able to
see that their job posting exists at the top of the list.