Build shell of submit page
The submit page itself doesn't do anything but we're collecting all of
the input data into the component state, so that we can submit it for a
mutation later on.

Something to consider is that we'll probably want to have the textareas
actually be markdown editors, so that people can format their posts.