Individual job posts are now loaded via Apollo
I've added a root query type for a single job with a numeric ID
parameter. Which means I was able to decouple the fact that we were
populating all of the ViewPage components using the jobs that were
queried on page load. This makes it significantly easier and much
cleaner code in terms of loading directly to a job post. It also lets us
simplify the query that's used on the HomePage, since we only need a
small excerpt of the information for displaying the rows.

Go GraphQL!